Refactor project structure: move UI components to UI_Components directory and update import paths
